Fix sqlite datetime handling for hot torrents
uploaded_at contains the ISO-8601 'T' separator, but sqlite uses string comparisons to filter the desired range. It's necessary to strip the 'T' by passing uploaded_at through the datetime function. Unfortunately this loses the use of the index. Alternatives might be to construct the datetimes manually to include the 'T' to match the uploaded_at field, or to insert uploaded_at as UTC to match sqlite's native datetime representation.

sqlite's default datetime zone is UTC, so the upper bound should not be modified to the localtime.